Implemented Functions to Generate a Stream of Jobs
Implemented a function to create a random job given ranges of parameters. By Default creates jobs to the same specifications of the paper. Also created a function to calculate the rate at which jobs should be randomly generated to yield the expected total utilization (70% according to paper). Math seems to check out and could be interesting for paper. Will need to test with simulations that it does actually yield the expected utilizations.