Created Instant Gratification Scheduler
Created new scheduler to be used as a baseline to compare our recurrent learning scheduler. It accepts the same parameters as the RL model, but rather than doing training models with the reward function, it simply calculates the immediate predicted reward function if it were to schedule on each cluster, then chooses the cluster that yields the maximum reward.